// src/store/cart/cart.types.ts
export enum CART_ACTION_TYPES {
    SET_CART_CLICKED = "SET_CART_CLICKED",
    ADD_CART_ITEM = "ADD_CART_ITEM",
    REMOVE_CART_ITEM = "REMOVE_CART_ITEM",
    ADD_CART_QUANTITY = "ADD_CART_QUANTITY",
    SUB_CART_QUANTITY = "SUB_CART_QUANTITY",
    CLEAR_CART = "CLEAR_CART"
}

// export const CART_ACTION_TYPES = {
//     SET_CART_CLICKED: 'SET_CART_CLICKED',
//     ADD_CART_ITEM: 'ADD_CART_ITEM',
//     REMOVE_CART_ITEM: 'REMOVE_CART_ITEM',
//     ADD_CART_QUANTITY: 'ADD_CART_QUANTITY',
//     SUB_CART_QUANTITY: 'SUB_CART_QUANTITY',
//     CLEAR_CART: 'CLEAR_CART'
// }
// export const CART_ACTION_TYPES = {
//     SET_CART_CLICKED: 'SET_CART_CLICKED',
//     ADD_CART_ITEM: 'ADD_CART_ITEM',
//     REMOVE_CART_ITEM: 'REMOVE_CART_ITEM',
//     ADD_CART_QUANTITY: 'ADD_CART_QUANTITY',
//     SUB_CART_QUANTITY: 'SUB_CART_QUANTITY',
//     CLEAR_CART: 'CLEAR_CART'
// }